+++ src/aui-data-editor/HISTORY.md
@@ -4,5 +4,6 @@
 
 ## @VERSION@
 
+* [AUI-1694](https://issues.liferay.com/browse/AUI-1694) Create grid field
 * [AUI-1689](https://issues.liferay.com/browse/AUI-1689) Create choice field
 * [AUI-1686](https://issues.liferay.com/browse/AUI-1686) Create editors for changing field settings